+++ examples/resources/dokploy_schedule/import.sh
@@ -1,0 +1,1 @@
+terraform import dokploy_schedule.prune s1a2b3c4d5e6f7g8h9i0j