+++ internal/provider/provider.go
@@ -24,6 +24,7 @@ import (
 	"github.com/vanillauys/terraform-provider-dokploy/internal/resources/environment"
 	"github.com/vanillauys/terraform-provider-dokploy/internal/resources/mount"
 	"github.com/vanillauys/terraform-provider-dokploy/internal/resources/project"
+	"github.com/vanillauys/terraform-provider-dokploy/internal/resources/schedule"
 )
 
 var _ provider.Provider = (*DokployProvider)(nil)
@@ -167,6 +168,7 @@ func (p *DokployProvider) Resources(_ context.Context) []func() resource.Resourc
 		domain.NewResource,
 		mount.NewResource,
 		destination.NewResource,
+		schedule.NewResource,
 		appchild.NewResource(appchild.PortKind()),
 		appchild.NewResource(appchild.RedirectKind()),
 		appchild.NewResource(appchild.SecurityKind()),